More Editing, New Barber Draft Site
Continued editing of the sites, I think forum is the only one still
needing editing (and a lot of it).

Added a Barber Draft Site, it's basically a site of images giving the
illusion of a website. Still needs to be "mobilised".

Thinking of adding a site section for the gooey image as that's a dead
image at the moment.