Updated examples main page + General corrections
Examples main page looks more visually appealing now, also contains more
information about the individual projects. moved the CSS to inline due
to it's small size, hopefully has improved load time.

Fixed the JS in the Toy company and a typo/

Attempted to adjust some of the spacing in the forum and moved the
copywrite message.